Move validator rest_api to its own crate
