+++ Cargo.toml
@@ -43,6 +43,7 @@ members = [
     "tests/node_test_rig",
     "lcli",
     "validator_client",
+    "validator_client/rest_api",
     "account_manager",
     "lighthouse",
     "lighthouse/environment"
