+++ validator_client/rest_api/src/errors.rs
