+++ validator_client/rest_api/src/response_builder.rs
