+++ validator_client/rest_api/src/status.rs
